feat(install): add webpacker generator
